+++ src/sensors/airdata_mgr.cxx
@@ -87,7 +87,7 @@ void AirData_init() {
     // traverse configured modules
     pyPropertyNode group_node = pyGetNode("/config/sensors/airdata_group", true);
     vector<string> children = group_node.getChildren();
-    printf("Found %d airdata sections\n", children.size());
+    printf("Found %d airdata sections\n", (int)children.size());
     for ( unsigned int i = 0; i < children.size(); i++ ) {
 	pyPropertyNode section = group_node.getChild(children[i].c_str());
 	sections.push_back(section);
