+++ src/sensors/gps_mgr.cxx
@@ -64,7 +64,7 @@ void GPS_init() {
     // traverse configured modules
     pyPropertyNode group_node = pyGetNode("/config/sensors/gps_group", true);
     vector<string> children = group_node.getChildren();
-    printf("Found %d gps sections\n", children.size());
+    printf("Found %d gps sections\n", (int)children.size());
     for ( unsigned int i = 0; i < children.size(); i++ ) {
 	pyPropertyNode section = group_node.getChild(children[i].c_str());
 	sections.push_back(section);
